test_case/failure/0063.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>
    [0063: Failed to search read a vss datapoint due to expired authorization token]
</title>

<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">
<link rel="stylesheet" type="text/css" href="/css/style-general.css">

<script src="/js/testharness.js"></script>
<script src="/js/testharnessreport.js"></script>
<script src="/js/vehicle-test-config.js"></script>
<script src="/js/vehicle-test-helper.js"></script>

<div id="test-title">
    [0063: Failed to search read a vss datapoint due to expired authorization token]
</div>
<br>
<div id="test-assertion">
    <b>Test assertion:</b><br>
    When the client makes getRequest to the authorization-needed server with filter containing "paths" op-type and correct op-value and expired authorization token,<br>
    [401 Token Expired] error returned.<br>
</div>
<div id="scenario">
    <br>
    <b>Test Scenario:</b><br>
    1. Send getRequest specifying an action, data path, requestId and expired JWT.<br>
    2. Confirm [401 Token Expired] error response.<br>
    <br>
    <b>Expected Result:</b><br>
    - Received [401 Token Expired] error response.<br>
</div>

<div id="result"></div>
<br>
<div id="log"></div>

<script>



    function sleep(time) {
        return new Promise((resolve) => setTimeout(resolve, time));
    }
    //https://stackoverflow.com/questions/951021/what-is-the-javascript-version-of-sleep

    function getResponseHeaderMap(xhr) {
        const headers = {};
        xhr.getAllResponseHeaders()
            .trim()
            .split(/[\r\n]+/)
            .map(value => value.split(/: /))
            .forEach(keyValue => {
                headers[keyValue[0].trim()] = keyValue[1].trim();
            });
        return headers;
    }

    function getAccessToken() {
        var xhttp = new XMLHttpRequest();
        xhttp.onreadystatechange = function () {
            if (this.readyState == 4 && this.status == 200) {
                var response_data = JSON.parse(this.response);
                access_token = response_data['access']
                var header = getResponseHeaderMap(this)
                console.log(this)
                console.log(header)

            }
        };
        // REST API 
        let request = "POST"
        let protocol = "http"
        let host = "127.0.0.1:8000"
        let path = "api/token"
        let url = protocol + "://" + host + "/" + path
        let data = new FormData();
        data.append('username', 'skku');
        data.append('password', 'skku');
        xhttp.open(request, url, false);
        xhttp.send(data);
    }

    function executeRestAPIWithAccessToken(access_token) {
        var xhttp = new XMLHttpRequest();
        xhttp.onreadystatechange = function () {
            if (this.readyState == 4 && this.status == 200) {
                var response_data = JSON.parse(this.response);
                var header = getResponseHeaderMap(this)
                var status_msg = "HTTP/1.1 " + this.status + " " + this.statusText
                var success = true
                console.log(this.getAllResponseHeaders())
                t.step_func(function () {
                    success = isGetSuccessResponseForHTTPS(response_data.data)
                    if (success) {
                        //assert_true(true, "Get response received. value = " + response_data.data.dp.value);
                        helper_terminate_success_for_https("Get response received. ", status_msg, header, response_data);
                    } else {
                        helper_terminate_failure("Get method failed");
                    }
                })();
            } else if (this.readyState == 4) {
                var response_data = JSON.parse(this.response)
                var header = getResponseHeaderMap(this)
                var status_msg = "HTTP/1.1 " + this.status + " " + this.statusText
                helper_terminate_failure_for_https("Get method failed ", status_msg, header, response_data);
            }
        };
        // REST API 
        let request = "GET"
        let protocol = "http"
        let host = "127.0.0.1:8000"
        let path = "Vehicle/Cabin/Door"
        let url = protocol + "://" + host + "/" + path
        let header = {}
        header['Authorization'] = 'Bearer ' + access_token
        let params = 'filter={"op-type":"paths", "op-value":"Row1/Left/IsOpen"}';
        if (params.length > 0) {
            console.log(params.length)
            url = url + "?" + params
            path = path + "?" + params
        }
        addLogRequestMessageForHTTPS(request + " /" + path + " HTTP/1.1", host, header)
        xhttp.open(request, url, true);
        for (let key in header) {
            xhttp.setRequestHeader(key, header[key]);
        }
        xhttp.send();
    }


    var t = async_test("Get success case");
    let access_token = ''
    getAccessToken()
    // executeRestAPIWithAccessToken(access_token)

    /*
    settings.py 수정한 후에 실행해야 error 반환

    SIMPLE_JWT = {
        'SIGNING_KEY': SECRET_KEY,
        'ALGORITHM': 'HS256',
        'ACCESS_TOKEN_LIFETIME': datetime.timedelta(seconds=1),
        'REFRESH_TOKEN_LIFETIME': datetime.timedelta(days=28),
    }
    */

    sleep(3000).then(() => {
        executeRestAPIWithAccessToken(access_token);
    });

</script>

</html>
